backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.1
pydantic>=2.9.2
python-dotenv>=1.0.1
aiofiles>=24.1.0
strands-agents-builder>=0.1.0
